Rename `getClipPath` file
