+++ src/components/clip-path-control/getClipPath.js
